+++ template/pie-ts/index.ts
@@ -2,8 +2,8 @@ import {makePie, MiraiPieApplication} from 'miraipie';
 
 module.exports = (ctx: MiraiPieApplication) => {
     ctx.pie(makePie({
-        id: 'my-pie',
-        name: 'my pie',
+        id: '{{id}}',
+        name: '{{name}}',
         version: '0.0.1',
         author: '',
         filters: [],