+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.1.10
+
+- 修复 日志系统事件问题。
+- 添加 等待下一条消息方法 `Chat.nextMessage(timeout: number)` 。
+- 优化 `pie` 和 `adapter` 生命周期钩子 `installed` 的执行时间。
+- 优化 CLI新建 `pie` 项目操作。
+
 ## v1.1.9
 
 - 修复 日志系统事件问题。